Android App Version: 1.0.1+2

+++ Android App/lib/handlers/http_request_handler.dart
@@ -1,5 +1,7 @@
 import 'dart:convert';
 import 'dart:io';
+import 'dart:typed_data';
+import 'package:connectivity_plus/connectivity_plus.dart';
 import 'package:flutter/material.dart';
 import '../build_config.dart';
 import '../screens/login.dart';
@@ -20,6 +22,17 @@ class HttpRequestHandler {
     };
   }
 
+  Future<bool> _checkNetwork() async {
+    final List<ConnectivityResult> connectivityResult = await (Connectivity().checkConnectivity());
+    if (connectivityResult.contains(ConnectivityResult.mobile)) {
+      return true;
+    } else if (connectivityResult.contains(ConnectivityResult.wifi)) {
+      return true;
+    }
+    NotificationHandler.showErrorNotification("No internet connection available.");
+    return false;
+  }
+
   void signOut() async {
     Navigator.of(buildContext!).pushReplacement(
       MaterialPageRoute(
@@ -29,24 +42,24 @@ class HttpRequestHandler {
   }
 
   signInRequest(String userName, String password) async {
+    if(!await _checkNetwork()){
+      return {};
+    }
     String signUrl = '${BuildConfig.serverUrl}/api/auth/login';
     var body = {
       "username": userName,
       "password": password,
     };
     try {
-      http.Response response = await http.post(
-          Uri.parse(signUrl),
-          headers: headers,
-          body: jsonEncode(body)
-      );
+      http.Response response = await http.post(Uri.parse(signUrl),
+          headers: headers, body: jsonEncode(body));
       Map<String, dynamic> respJson = jsonDecode(response.body);
-      if(response.statusCode == 200){
+      if (response.statusCode == 200) {
         respJson['status'] = 200;
         BuildConfig.authorization = respJson['token'];
         BuildConfig.username = respJson['username'];
         return respJson;
-      }else{
+      } else {
         NotificationHandler.showErrorNotification(respJson["message"]);
         return respJson;
       }
@@ -57,8 +70,12 @@ class HttpRequestHandler {
     }
   }
 
-  Future<Map<String, dynamic>> transactionGet(String dateOrMonth) async {
-    String transactionUrl = '${BuildConfig.serverUrl}/api/dashboard/summary?dateOrMonth=$dateOrMonth';
+  Future<Map<String, dynamic>> transactionGet(String dateOrMonth, String party) async {
+    if(!await _checkNetwork()){
+      return {};
+    }
+    String transactionUrl =
+        '${BuildConfig.serverUrl}/api/dashboard/summary?dateOrMonth=$dateOrMonth&party=$party';
     try {
       http.Response response = await http.get(
         Uri.parse(transactionUrl),
@@ -68,10 +85,11 @@ class HttpRequestHandler {
       if (response.statusCode == 200) {
         respJson['status'] = 200;
         return respJson;
-      }else if(response.statusCode == 401){
+      } else if (response.statusCode == 401) {
         signOut();
         return {};
       } else {
+        NotificationHandler.showErrorNotification(respJson["message"]);
         return respJson;
       }
     } catch (e) {
@@ -80,8 +98,12 @@ class HttpRequestHandler {
     }
   }
 
-  Future<Map<String, dynamic>> monthlyTransactionGet(String dateOrMonth) async {
-    String transactionUrl = '${BuildConfig.serverUrl}/api/dashboard/summary/grouped?startDate=$dateOrMonth';
+  Future<Map<String, dynamic>> monthlyTransactionGet(String dateOrMonth, String party) async {
+    if(!await _checkNetwork()){
+      return {};
+    }
+    String transactionUrl =
+        '${BuildConfig.serverUrl}/api/dashboard/summary/grouped?startDate=$dateOrMonth&party=$party';
     try {
       http.Response response = await http.get(
         Uri.parse(transactionUrl),
@@ -91,10 +113,11 @@ class HttpRequestHandler {
       if (response.statusCode == 200) {
         respJson['status'] = 200;
         return respJson;
-      } else if(response.statusCode == 401){
+      } else if (response.statusCode == 401) {
         signOut();
         return {};
       } else {
+        NotificationHandler.showErrorNotification(respJson["message"]);
         return respJson;
       }
     } catch (e) {
@@ -103,8 +126,13 @@ class HttpRequestHandler {
     }
   }
 
-  Future<Map<String, dynamic>> dateRangeTransactionGet(String startDate, String endDate) async {
-    String transactionUrl = '${BuildConfig.serverUrl}/api/dashboard/summary/grouped?startDate=$startDate&endDate=$endDate';
+  Future<Map<String, dynamic>> dateRangeTransactionGet(
+      String startDate, String endDate, String party) async {
+    if(!await _checkNetwork()){
+      return {};
+    }
+    String transactionUrl =
+        '${BuildConfig.serverUrl}/api/dashboard/summary/grouped?startDate=$startDate&endDate=$endDate&party=$party';
     try {
       http.Response response = await http.get(
         Uri.parse(transactionUrl),
@@ -114,10 +142,11 @@ class HttpRequestHandler {
       if (response.statusCode == 200) {
         respJson['status'] = 200;
         return respJson;
-      }else if(response.statusCode == 401){
+      } else if (response.statusCode == 401) {
         signOut();
         return {};
       } else {
+        NotificationHandler.showErrorNotification(respJson["message"]);
         return respJson;
       }
     } catch (e) {
@@ -127,6 +156,9 @@ class HttpRequestHandler {
   }
 
   fileUpload(File file) async {
+    if(!await _checkNetwork()){
+      return {};
+    }
     String url = '${BuildConfig.serverUrl}/api/file/upload';
     try {
       var uri = Uri.parse(url);
@@ -142,7 +174,7 @@ class HttpRequestHandler {
       var jsonResponse = json.decode(responseString);
       if (response.statusCode == 200) {
         return jsonResponse;
-      }else if(response.statusCode == 401){
+      } else if (response.statusCode == 401) {
         signOut();
       } else {
         return null;
@@ -154,23 +186,21 @@ class HttpRequestHandler {
   }
 
   postTransaction(var body) async {
+    if(!await _checkNetwork()){
+      return {};
+    }
     String url = '${BuildConfig.serverUrl}/api/transaction';
     try {
-      http.Response response = await http.post(
-          Uri.parse(url),
-          headers: headers,
-          body: jsonEncode(body)
-      );
+      http.Response response = await http.post(Uri.parse(url),
+          headers: headers, body: jsonEncode(body));
       Map<String, dynamic> respJson = jsonDecode(response.body);
-      if(response.statusCode == 201){
+      if (response.statusCode == 201) {
         respJson['status'] = 201;
         return respJson;
-      }else if(response.statusCode == 401){
+      } else if (response.statusCode == 401) {
         signOut();
-      }else if(response.statusCode == 401){
-        signOut();
-      }
-      else{
+      }else {
+        NotificationHandler.showErrorNotification(respJson["message"]);
         return respJson;
       }
     } catch (e) {
@@ -180,20 +210,21 @@ class HttpRequestHandler {
   }
 
   updateTransaction(var body, String id) async {
+    if(!await _checkNetwork()){
+      return {};
+    }
     String url = '${BuildConfig.serverUrl}/api/transaction/$id';
     try {
-      http.Response response = await http.put(
-          Uri.parse(url),
-          headers: headers,
-          body: jsonEncode(body)
-      );
+      http.Response response = await http.put(Uri.parse(url),
+          headers: headers, body: jsonEncode(body));
       Map<String, dynamic> respJson = jsonDecode(response.body);
-      if(response.statusCode == 200){
+      if (response.statusCode == 200) {
         respJson['status'] = 200;
         return respJson;
-      }else if(response.statusCode == 401){
+      } else if (response.statusCode == 401) {
         signOut();
-      }else{
+      } else {
+        NotificationHandler.showErrorNotification(respJson["message"]);
         return respJson;
       }
     } catch (e) {
@@ -203,17 +234,20 @@ class HttpRequestHandler {
   }
 
   deleteTransaction(String id) async {
+    if(!await _checkNetwork()){
+      return {};
+    }
     String url = '${BuildConfig.serverUrl}/api/transaction/$id';
     try {
       http.Response response = await http.delete(
-          Uri.parse(url),
-          headers: headers,
+        Uri.parse(url),
+        headers: headers,
       );
-      if(response.statusCode == 204){
+      if (response.statusCode == 204) {
         return true;
-      }else if(response.statusCode == 401){
+      } else if (response.statusCode == 401) {
         signOut();
-      }else{
+      } else {
         return false;
       }
     } catch (e) {
@@ -223,6 +257,9 @@ class HttpRequestHandler {
   }
 
   getParty() async {
+    if(!await _checkNetwork()){
+      return [];
+    }
     String url = '${BuildConfig.serverUrl}/api/transaction/party';
     try {
       http.Response response = await http.get(
@@ -232,7 +269,7 @@ class HttpRequestHandler {
       if (response.statusCode == 200) {
         List<dynamic> respJson = jsonDecode(response.body);
         return respJson;
-      }else if(response.statusCode == 401){
+      } else if (response.statusCode == 401) {
         signOut();
       } else {
         return [];
@@ -244,6 +281,9 @@ class HttpRequestHandler {
   }
 
   Future<List<dynamic>> getHistory() async {
+    if(!await _checkNetwork()){
+      return [];
+    }
     String url = '${BuildConfig.serverUrl}/api/history';
     try {
       http.Response response = await http.get(
@@ -253,7 +293,7 @@ class HttpRequestHandler {
       if (response.statusCode == 200) {
         List<dynamic> respJson = jsonDecode(response.body);
         return respJson;
-      }else if(response.statusCode == 401){
+      } else if (response.statusCode == 401) {
         signOut();
         return [];
       } else {
@@ -266,19 +306,24 @@ class HttpRequestHandler {
   }
 
   Future<Map<String, dynamic>> fileVaultGet(String folderId) async {
-    String url = '${BuildConfig.serverUrl}/api/vault/folders?folderId=$folderId';
+    if(!await _checkNetwork()){
+      return {};
+    }
+    String url =
+        '${BuildConfig.serverUrl}/api/vault/folders?folderId=$folderId';
     try {
       http.Response response = await http.get(
         Uri.parse(url),
         headers: headers,
       );
+      Map<String, dynamic> respJson = jsonDecode(response.body);
       if (response.statusCode == 200) {
-        Map<String, dynamic> respJson = jsonDecode(response.body);
         return respJson;
-      }else if(response.statusCode == 401){
+      } else if (response.statusCode == 401) {
         signOut();
         return {};
       } else {
+        NotificationHandler.showErrorNotification(respJson["message"]);
         return {};
       }
     } catch (e) {
@@ -288,20 +333,22 @@ class HttpRequestHandler {
   }
 
   addFolder(var body) async {
-    String url = '${BuildConfig.serverUrl}/api/vault/folders?name=${body['name']}&parentId=${body['parentId']}';
+    if(!await _checkNetwork()){
+      return {};
+    }
+    String url =
+        '${BuildConfig.serverUrl}/api/vault/folders?name=${body['name']}&parentId=${body['parentId']}';
     try {
-      http.Response response = await http.post(
-          Uri.parse(url),
-          headers: headers,
-          body: jsonEncode(body)
-      );
+      http.Response response = await http.post(Uri.parse(url),
+          headers: headers, body: jsonEncode(body));
       Map<String, dynamic> respJson = jsonDecode(response.body);
-      if(response.statusCode == 200){
+      if (response.statusCode == 200) {
         respJson['status'] = 200;
         return respJson;
-      }else if(response.statusCode == 401){
+      } else if (response.statusCode == 401) {
         signOut();
-      }else{
+      } else {
+        NotificationHandler.showErrorNotification(respJson["message"]);
         return respJson;
       }
     } catch (e) {
@@ -311,6 +358,9 @@ class HttpRequestHandler {
   }
 
   vaultFileUpload(File file, String folderId) async {
+    if(!await _checkNetwork()){
+      return {};
+    }
     String url = '${BuildConfig.serverUrl}/api/vault/files';
     try {
       var uri = Uri.parse(url);
@@ -327,9 +377,10 @@ class HttpRequestHandler {
       var jsonResponse = json.decode(responseString);
       if (response.statusCode == 200) {
         return jsonResponse;
-      }else if(response.statusCode == 401){
+      } else if (response.statusCode == 401) {
         signOut();
       } else {
+        NotificationHandler.showErrorNotification(jsonResponse["message"]);
         return null;
       }
     } catch (e) {
@@ -339,17 +390,20 @@ class HttpRequestHandler {
   }
 
   deleteFiles(String id) async {
+    if(!await _checkNetwork()){
+      return {};
+    }
     String url = '${BuildConfig.serverUrl}/api/vault/files/$id';
     try {
       http.Response response = await http.delete(
         Uri.parse(url),
         headers: headers,
       );
-      if(response.statusCode == 204){
+      if (response.statusCode == 204) {
         return true;
-      }else if(response.statusCode == 401){
+      } else if (response.statusCode == 401) {
         signOut();
-      }else{
+      } else {
         return false;
       }
     } catch (e) {
@@ -359,23 +413,62 @@ class HttpRequestHandler {
   }
 
   deleteFolder(String id) async {
+    if(!await _checkNetwork()){
+      return {};
+    }
     String url = '${BuildConfig.serverUrl}/api/vault/folders/$id';
     try {
       http.Response response = await http.delete(
         Uri.parse(url),
         headers: headers,
       );
-      if(response.statusCode == 204){
+      if (response.statusCode == 204) {
         return true;
-      }else if(response.statusCode == 401){
+      } else if (response.statusCode == 401) {
         signOut();
         return {};
-      }else{
+      } else {
         return false;
       }
     } catch (e) {
       print("Error in deleteFolder: $e");
       return false;
+    }
+  }
+
+  Future<Uint8List> fetchBlob(String fileUuid) async {
+    if(!await _checkNetwork()){
+      throw Exception("Failed to load blob data");
+    }
+    String url = '${BuildConfig.serverUrl}/api/file/download/$fileUuid';
+    final response = await http.get(
+      Uri.parse(url),
+      headers: headers,
+    );
+    print(response.statusCode);
+    if (response.statusCode == 200) {
+      return response.bodyBytes; // Blob data
+    } else {
+      NotificationHandler.showErrorNotification(json.decode(response.body)["message"]);
+      throw Exception("Failed to load blob data");
+    }
+  }
+
+  Future<Uint8List> fetchVaultBlob(String fileUuid) async {
+    if(!await _checkNetwork()){
+      throw Exception("Failed to load blob data");
+    }
+    String url = '${BuildConfig.serverUrl}/api/vault/files/download/$fileUuid';
+    final response = await http.get(
+      Uri.parse(url),
+      headers: headers,
+    );
+    print(response.statusCode);
+    if (response.statusCode == 200) {
+      return response.bodyBytes; // Blob data
+    } else {
+      NotificationHandler.showErrorNotification(json.decode(response.body)["message"]);
+      throw Exception("Failed to load blob data");
     }
   }
 }
